+++ src/modules/emails/emails.service.ts
@@ -61,9 +61,9 @@ export class EmailsService {
     const email = await this.prisma.emails.findOne({
       where: { id },
     });
-    if (email.userId !== userId) throw new UnauthorizedException();
     if (!email)
       throw new HttpException('Email not found', HttpStatus.NOT_FOUND);
+    if (email.userId !== userId) throw new UnauthorizedException();
     return this.prisma.expose<emails>(email);
   }
 
@@ -71,6 +71,8 @@ export class EmailsService {
     const testEmail = await this.prisma.emails.findOne({
       where: { id },
     });
+    if (!testEmail)
+      throw new HttpException('Email not found', HttpStatus.NOT_FOUND);
     if (testEmail.userId !== userId) throw new UnauthorizedException();
     const email = await this.prisma.emails.delete({
       where: { id },
